+++ src/components/finutils/src/common/dev/mod.rs
@@ -9,18 +9,16 @@
 
 mod init;
 
+use super::ddev::Ports;
 use chaindev::tm_dev::{
-    self, CustomOp, EnvMeta, EnvName, EnvOpts, Node, NodeOptsGenerator, NodePorts, Op,
+    self, CustomOps, EnvMeta, EnvName, EnvOpts, Node, NodeOptsGenerator, Op,
 };
-use ledger::staking::{
-    td_addr_to_bytes, Validator as StakingValidator, ValidatorKind, FRA, VALIDATORS_MIN,
-};
-use noah::xfr::sig::XfrKeyPair;
+use init::{BankAccount, InitialValidator};
 use rucv3::*;
 use serde::{Deserialize, Serialize};
 use std::{fmt::Write, thread};
 
-#[derive(Debug)]
+#[derive(Debug, Default)]
 pub struct EnvCfg {
     // the name of this env
     pub name: EnvName,
@@ -28,57 +26,6 @@ pub struct EnvCfg {
     // which operation to trigger,
     // default value: `Ops::Show`
     pub ops: Ops,
-
-    // seconds between two blocks,
-    // default value: 3
-    pub block_itv_secs: f32,
-
-    // how many initial validators should be created
-    pub initial_validator_num: u8,
-
-    // initialized once in `Ops::Create`,
-    // default value: "127.0.0.1"
-    pub host_ip: Option<String>,
-
-    // specify this option if you want to use a custom version of abcid
-    pub abcid_bin: Option<String>,
-
-    // specify this option if you want to use a custom version of tendermint
-    pub tendermint_bin: Option<String>,
-
-    // only used in `Ops::Create`
-    pub abcid_extra_flags: Option<String>,
-
-    // only used in `Ops::Create`
-    pub tendermint_extra_flags: Option<String>,
-
-    pub force_create: bool,
-
-    // default value: 2152
-    pub evm_chain_id: u64,
-
-    // only used in `Ops::Create`
-    // used in `Ops::Create`
-    pub checkpoint_file: Option<String>,
-}
-
-impl Default for EnvCfg {
-    fn default() -> Self {
-        Self {
-            name: EnvName::default(),
-            ops: Ops::default(),
-            block_itv_secs: 3.0,
-            initial_validator_num: VALIDATORS_MIN as u8,
-            host_ip: None,
-            abcid_bin: None,
-            tendermint_bin: None,
-            abcid_extra_flags: None,
-            tendermint_extra_flags: None,
-            force_create: false,
-            evm_chain_id: 2152,
-            checkpoint_file: None,
-        }
-    }
 }
 
 impl From<EnvCfg> for tm_dev::EnvCfg<(), CustomData, Ports, InitOps> {
@@ -196,64 +143,6 @@ impl Default for Ops {
     }
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct Ports {
-    #[serde(rename = "web3_http_service")]
-    web3_http: u16,
-    #[serde(rename = "web3_websocket_service")]
-    web3_ws: u16,
-    #[serde(rename = "abcid_ledger_query_service")]
-    app_8668: u16,
-    #[serde(rename = "abcid_submission_service")]
-    app_8669: u16,
-    #[serde(rename = "tendermint_p2p_service")]
-    tm_p2p: u16,
-    #[serde(rename = "tendermint_rpc_service")]
-    tm_rpc: u16,
-    #[serde(rename = "abcid_abci_service")]
-    app_abci: u16,
-}
-
-impl NodePorts for Ports {
-    fn app_reserved() -> Vec<u16> {
-        vec![8545, 8546, 8668, 8669]
-    }
-    fn try_create(ports: &[u16]) -> Result<Self> {
-        if ports.len() != Self::reserved().len() {
-            return Err(eg!("invalid length"));
-        }
-        Ok(Self {
-            web3_http: ports[0],
-            web3_ws: ports[1],
-            app_8668: ports[2],
-            app_8669: ports[3],
-            tm_p2p: ports[4],
-            tm_rpc: ports[5],
-            app_abci: ports[6],
-        })
-    }
-    fn get_port_list(&self) -> Vec<u16> {
-        vec![
-            self.web3_http,
-            self.web3_ws,
-            self.app_8668,
-            self.app_8669,
-            self.tm_p2p,
-            self.tm_rpc,
-            self.app_abci,
-        ]
-    }
-    fn get_sys_p2p(&self) -> u16 {
-        self.tm_p2p
-    }
-    fn get_sys_rpc(&self) -> u16 {
-        self.tm_rpc
-    }
-    fn get_sys_abci(&self) -> u16 {
-        self.app_abci
-    }
-}
-
 #[derive(Clone, Debug, Serialize, Deserialize)]
 struct OptsGenerator;
 
@@ -319,7 +208,7 @@ enum InitOps {
 
 type Env = EnvMeta<CustomData, Node<Ports>>;
 
-impl CustomOp for InitOps {
+impl CustomOps for InitOps {
     fn exec(&self, env_name: &EnvName) -> Result<()> {
         match self {
             InitOps::Init => Env::load_env_by_name(env_name)
@@ -376,58 +265,4 @@ struct CustomData {
 
     #[serde(rename = "initial_pos_settings")]
     initial_validators: Vec<InitialValidator>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct BankAccount {
-    wallet_address: String,
-    public_key: String,
-    secret_key: String,
-    mnemonic_words: String,
-}
-
-impl BankAccount {
-    const BANK_ACCOUNT_ADDR: &str =
-        "fra18xkez3fum44jq0zhvwq380rfme7u624cccn3z56fjeex6uuhpq6qv9e4g5";
-    const BANK_ACCOUNT_PUBKEY: &str = "Oa2RRTzdayA8V2OBE7xp3n3NKrjGJxFTSZZybXOXCDQ=";
-    const BANK_ACCOUNT_SECKEY: &str = "Ew9fMaryTL44ZXnEhcF7hQ-AB-fxgaC8vyCH-hCGtzg=";
-    const BANK_ACCOUNT_MNEMONIC: &str = "field ranch pencil chest effort coyote april move injury illegal forest amount bid sound mixture use second pet embrace twice total essay valve loan";
-}
-
-impl Default for BankAccount {
-    fn default() -> Self {
-        Self {
-            wallet_address: Self::BANK_ACCOUNT_ADDR.to_owned(),
-            public_key: Self::BANK_ACCOUNT_PUBKEY.to_owned(),
-            secret_key: Self::BANK_ACCOUNT_SECKEY.to_owned(),
-            mnemonic_words: Self::BANK_ACCOUNT_MNEMONIC.to_owned(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct InitialValidator {
-    tendermint_addr: String,
-    tendermint_pubkey: String,
-
-    xfr_keypair: XfrKeyPair,
-    xfr_mnemonic: String,
-    xfr_wallet_addr: String,
-}
-
-impl From<&InitialValidator> for StakingValidator {
-    fn from(v: &InitialValidator) -> StakingValidator {
-        StakingValidator {
-            td_pubkey: base64::decode(&v.tendermint_pubkey).unwrap(),
-            td_addr: td_addr_to_bytes(&v.tendermint_addr).unwrap(),
-            td_power: 400_0000 * FRA,
-            commission_rate: [1, 100],
-            id: v.xfr_keypair.get_pk(),
-            memo: Default::default(),
-            kind: ValidatorKind::Initiator,
-            signed_last_block: false,
-            signed_cnt: 0,
-            delegators: Default::default(),
-        }
-    }
 }
